Streamlined configuration
Config Constructors should now be with the field definition, or whatever the technical term for that is. Loading and saving of configuration moved to an extension function "ReloadConfigurable", which is now called by ModuleHandler. In short, Module : IConfigurable shall no longer call Save or Load.

Oh also IConfigurable now takes in a generic that defines which type of configurable it is.